Prefer explicit data or tested parsers over regex
Regexes are commonly regarded as hard to read and to understand. Of the
three regexes used in recapupload, two of them can be replaced. The
first tries to parse data from a JSON field that is already available to
us in another field in the same object. The second tries to use regex to
parse a URL, but we can use the already imported urlparse library to do
the same.

The third regex is actually the most complex and parses XML, which is
a [Bad Thing™](https://stackoverflow.com/a/1732454). There are some
[tradeoffs](https://docs.python.org/2/library/xml.html#xml-vulnerabilities)
to using Python's built-in XML parsers with external data, so we will
punt the decision about what to do here for later.